+++ zallet/src/commands.rs
@@ -29,6 +29,8 @@ mod start;
 #[cfg(zallet_build = "wallet")]
 mod export_mnemonic;
 #[cfg(zallet_build = "wallet")]
+mod generate_encryption_identity;
+#[cfg(zallet_build = "wallet")]
 mod generate_mnemonic;
 #[cfg(zallet_build = "wallet")]
 mod import_mnemonic;